build cleanly against AOSP 12 toolchain + boot in TWRP
Patches on top of Lenovo's OSS kernel drop (imported from
T4rp/Lenovo-TB-8505FS-Kernel) to make this 4.9.117 tree build against
the AOSP 12 host toolchain and actually boot to TWRP on the TB-8505F.

- tools/dct/*.py: converted Python 2 -> Python 3. MTK's DrvGen runs at
  build time; it used print-as-statement, cmp(), string.atoi(), etc.
  Added a 'def cmp' shim where needed; switched configparser to
  strict=False since the .cmp files have duplicate keys; renamed local
  'list' variables that shadowed the builtin once 2to3 introduced
  list(map.items()) calls.

- Makefile: KBUILD_CFLAGS += -Wno-error plus specific -Wno-error=<class>
  lines for maybe-uninitialized, format=, format-security,
  unused-variable/function/but-set-variable. GCC 4.9 fires on things
  clang silently accepts; -Werror makes them fatal.

- sound/soc/mediatek/Makefile: stripped -Werror and
  -Wno-incompatible-pointer-types from subdir-ccflags-y. The latter is
  unknown to GCC 4.9 and fatal under -Werror.

- s5k4h7{,_qtech}_mipi_raw/*.c: pow1/pow2 declared double but the
  kernel can't touch FP registers (-mgeneral-regs-only). Bodies were
  already broken (c = a*b in a loop, returning 2*j), so we just rewrite
  the signatures + callers to int and move on.

- ram_console/mtk_ram_console.c: NULL-guard aee_rr_rec_*_pa() in the
  three accessors, and move ram_console_buffer_pa assignment inside the
  ioremap_wc success branch. If the DTB lacks /chosen/ram_console the
  old code left _pa set but _buffer (virtual) NULL, and RR_LINUX_PA
  macro dereferences both -> NULL deref at fiq_cache_init.

- cpuidle_v3/mtk_cpuidle.c: wrap cpuidle_fp() and cpuidle_fp_reset() in
  if (cpuidle_fp_va). Same failure mode as above: if ram_console didn't
  map, cpuidle_fp_va is NULL and secondary CPUs entering idle for the
  first time crash at offset cpu*4.

- arch/arm64/configs/akita_row_wifi_defconfig: disable
  CONFIG_SND_SOC_AC107 and CONFIG_INPUT_SAR_SX9311. Both need
  compiler-specific workarounds we don't need for TWRP.

See README.prism.md for details. Tested: boots TWRP on hardware.

+++ drivers/misc/mediatek/imgsensor/src/common/v1/s5k4h7_mipi_raw/s5k4h7mipiraw_Sensor.c
@@ -1139,9 +1139,9 @@ static void slim_video_setting(void)
 //#define s5k4h7yxFRONT_I2CBUS    (4)
 
 #if S5K4H7_OTP
-double pow1(double a,int b)
+int pow1(int a,int b)
 {
-	double c=0.0;
+	int c=0;
 	int i = 0;
 	for (i=1;i<b;i++){
 		c = a*b;
@@ -1159,8 +1159,8 @@ int GetCRC(unsigned char *Data, unsigned int checksum_start,unsigned int Len)
         int hexonly = Data[checksum_start + i];
         for (j=7; j>=0; j--)
         {
-            DoInvert = hexonly & ((int)pow1(2.0,j));
-            DoInvert = DoInvert / (pow1(2.0,j));
+            DoInvert = hexonly & (pow1(2,j));
+            DoInvert = DoInvert / (pow1(2,j));
             DoInvert = DoInvert ^ CRC[15];
 
             CRC[15] = CRC[14] ^ DoInvert;
@@ -1184,7 +1184,7 @@ int GetCRC(unsigned char *Data, unsigned int checksum_start,unsigned int Len)
 
     for (i=0; i<16; i++)
     {
-        mycrc16x = mycrc16x + (CRC[i]*pow1(2.0,i));
+        mycrc16x = mycrc16x + (CRC[i]*pow1(2,i));
     }
     return mycrc16x;
 }
